Changes optional to ? in user repository. Fixed issues. Fixed tests.

+++ src/main/kotlin/ai/logsight/backend/logs/ports/web/responses/LogsReceiptResponse.kt
@@ -5,6 +5,6 @@ import java.util.*
 data class LogsReceiptResponse(
     val appId: UUID,
     val orderId: Long,
-    val logsCount: Long,
+    val logsCount: Int,
     val source: String,
 )
